ATO-146: Move `AuthenticationUserInfo` and `AuthenticationUserInfoStorageService` to shared To be used by the new lambda, these files must be in shared
