+++ integration-tests/src/test/java/uk/gov/di/authentication/services/AuthenticationUserInfoStorageServiceIntegrationTest.java
@@ -4,7 +4,7 @@ import com.nimbusds.oauth2.sdk.id.Subject;
 import com.nimbusds.openid.connect.sdk.claims.UserInfo;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
-import uk.gov.di.authentication.oidc.entity.AuthenticationUserInfo;
+import uk.gov.di.authentication.shared.entity.AuthenticationUserInfo;
 import uk.gov.di.authentication.sharedtest.extensions.AuthenticationCallbackUserInfoStoreExtension;
 
 import java.util.Optional;
