+++ shared-test/src/main/java/uk/gov/di/authentication/sharedtest/extensions/AuthenticationCallbackUserInfoStoreExtension.java
@@ -9,8 +9,8 @@ import software.amazon.awssdk.services.dynamodb.model.CreateTableRequest;
 import software.amazon.awssdk.services.dynamodb.model.KeySchemaElement;
 import software.amazon.awssdk.services.dynamodb.model.KeyType;
 import software.amazon.awssdk.services.dynamodb.model.ScalarAttributeType;
-import uk.gov.di.authentication.oidc.entity.AuthenticationUserInfo;
-import uk.gov.di.authentication.oidc.services.AuthenticationUserInfoStorageService;
+import uk.gov.di.authentication.shared.entity.AuthenticationUserInfo;
+import uk.gov.di.authentication.shared.services.AuthenticationUserInfoStorageService;
 import uk.gov.di.authentication.shared.services.ConfigurationService;
 import uk.gov.di.authentication.sharedtest.basetest.DynamoTestConfiguration;
